HTML updated
I updated the HTML so it looks better. I included any changes you guys can make as comments. All that is left is changing the website for looks and meet the accessibility guidelines. Make sure the images are in a folder called "images". It easier to just make a folder for all the html and css and a folder within for images. 

-Anil